django/requirements.txt
uvicorn[standard]==0.14.0
gunicorn==20.1.0
Django==4.1.8
channels==3.0.5
channels-redis==3.4.1
django-bootstrap5==22.1
django-q==1.3.9
django-environ==0.9.0
paho-mqtt==1.6.1
